requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
fonttools==4.45.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
Pillow==10.1.0
py-request-nacos==0.0.11
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
